refactor: unify /src/data and /src/content into single /src/content structure
MIGRATION COMPLETED:
- Moved all 10 files from /src/data/ to /src/content/ (preserving git history)
- Updated all import paths from @/data/* to @/content/*
- Updated relative imports from ../data/* to ../content/*
- Organized unified content structure with logical grouping:
  * Service data: collaboration.json, teaching.json, lessons.json, performance.json
  * Configuration: serviceConfiguration.json, ui-config.json, navigation.json
  * Social/Media: instagram-posts.json, testimonials.json
  * Contact/Gallery: contact.json, performance-gallery.json
  * Cross-service: crossServiceSuggestions.json
  * UI: ui/cta.json

VERIFIED FUNCTIONALITY:
- TypeScript compilation successful
- Production build successful (1.28s)
- All imports resolve correctly
- No broken references

BENEFITS:
- Single coherent content/data structure
- Cleaner, more intuitive architecture
- Better separation between content and application code
- Maintained file history with git mv

FILES UPDATED:
- 9 import statements updated across components, hooks, utils, services
- 1 UI subfolder created for component-specific configurations
- 10 JSON files successfully migrated
